jwt token login refactor

+++ apps/backend/src/app/absence/absence.service.ts
@@ -1,3 +1,4 @@
+import { JwtService } from '@nestjs/jwt';
 import { AbsenceDto } from './absence.dto';
 import { AbsenceEntity } from './absence.entity';
 import { HttpException, HttpStatus, Injectable } from '@nestjs/common';
@@ -9,39 +10,33 @@ Injectable();
 export class AbsenceService {
   constructor(
     @InjectRepository(AbsenceEntity)
-    private readonly absenceRepository: Repository<AbsenceEntity>
+    private readonly absenceRepository: Repository<AbsenceEntity>,
+    private readonly jwtService: JwtService
   ) {}
 
-  public getAbsences(id: number): Promise<AbsenceEntity[]> {
+  public async getAbsences(token: string): Promise<AbsenceEntity[]> {
     try {
-      return this.absenceRepository.find({ where: { user_id: id } });
+      const data = await this.jwtService.verify(token)
+      return await this.absenceRepository.find({ where: {userId: data.userId} })
     } catch {
-      throw new HttpException('Помилка запиті записів', HttpStatus.BAD_REQUEST);
+      throw new HttpException('Помилка при запиті записів', HttpStatus.BAD_REQUEST);
     }
   }
 
-  public async deleteAbsence(id: number): Promise<DeleteResult> {
+  public async deleteAbsence(id: number, token: string): Promise<DeleteResult> {
     try {
-      return await this.absenceRepository.delete({ id: id });
+      const data = this.jwtService.verify(token)
+      return await this.absenceRepository.delete({ id: id, userId: data.userId });
     } catch {
       throw new HttpException('Помилка при видаленні запису', HttpStatus.BAD_REQUEST);
     }
   }
 
-  public async createAbsence(absence: AbsenceDto): Promise<AbsenceEntity> {
-    let absenceEntity: AbsenceEntity = new AbsenceEntity();
-
-    absenceEntity.id = absence.id;
-    absenceEntity.user_id = absence.user_id;
-    absenceEntity.start = moment(absence.start).toDate();
-    absenceEntity.end = moment(absence.end).toDate();
-    absenceEntity.comment = absence.comment;
-    absenceEntity.type = absence.type;
-
+  public async createAbsence(absence: AbsenceDto, token: string): Promise<AbsenceEntity> {
     try {
+      const data = this.jwtService.verify(token)
       return await this.absenceRepository.save({
-        id: absence.id,
-        user_id: absence.user_id,
+        userId: Number(data.userId),
         start: moment(absence.start).toDate(),
         end: moment(absence.end).toDate(),
         comment: absence.comment,
@@ -52,11 +47,13 @@ export class AbsenceService {
     }
   }
 
-  public async editAbsence(absence: AbsenceDto): Promise<AbsenceEntity> {
+  public async editAbsence(absence: AbsenceDto, token: string): Promise<AbsenceEntity> {
     try {
+      const data = this.jwtService.verify(token)
+      const user = await this.absenceRepository.findOne( {where: {userId: data.userId}} )
+
       return await this.absenceRepository.save({
-        id: absence.id,
-        user_id: absence.user_id,
+        ...user,
         start: moment(absence.start).toDate(),
         end: moment(absence.end).toDate(),
         comment: absence.comment,
